core/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
httpretty==1.1.4
idna==3.10
model-bakery==1.20.5
PyJWT==2.9.0
python-dotenv==1.1.1
requests==2.32.4
ruff==0.12.1
sqlparse==0.5.3
urllib3==2.4.0
uv==0.7.12
